python/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
flask>=2.0.0
flask-cors>=3.0.0
